vtenc: fix pts / dts offset margin
Warning: this is a workaround. A proper fix for this will be addressed
upstream.

Using main or high profiles with B-frames, 2 frames can be reordered while
current code was assuming this could only happen with HEVC. The consequence was
that h264parse would run into situations where DTS < PTS and would override the
PTS with the DTS resulting in wrong timestamps. Since vtenc doesn't include
timing information, this caused frame drops on the guest resulting in poor
framerate and video quality.